requirements.txt
prefect>=2.0.0
httpx>=0.24.0
python-dotenv>=0.19.0
pandas>=1.3.0
matplotlib>=3.5.0
psycopg2-binary>=2.9.0
seaborn>=0.11.0
